+++ Pulp/Common/Threading/win32/Signal.h
@@ -9,7 +9,7 @@ X_NAMESPACE_BEGIN(core)
 class Signal
 {
 public:
-	static const uint32_t WAIT_INFINITE = -1;
+	static const uint32_t WAIT_INFINITE = static_cast<uint32_t>(-1);
 
 public:
 	Signal();
